% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouseConnector.R
\name{connect_clickhouse}
\alias{connect_clickhouse}
\title{Connect to ClickHouse over HTTP.}
\usage{
connect_clickhouse(
  host = "localhost",
  port = 8123,
  dbname = "default",
  user = "default",
  password = "",
  https = FALSE,
  ...
)
}
\arguments{
\item{host}{ClickHouse host name or IP address.}

\item{port}{ClickHouse HTTP port.}

\item{dbname}{Default ClickHouse database.}

\item{user}{ClickHouse user name.}

\item{password}{ClickHouse password.}

\item{https}{Use HTTPS instead of HTTP.}

\item{...}{Reserved for future compatibility; currently ignored.}
}
\value{
An object of class `gitariki_clickhouse_conn`.
}
\description{
Creates a lightweight stateless connection object used by the package's
ClickHouse helpers. The connection sends SQL through the ClickHouse HTTP
interface and does not keep an open DBI session.
}
\examples{
conn <- connect_clickhouse(host = "localhost", port = 8123)
}
