% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSV-database.R
\name{download_osv_database}
\alias{download_osv_database}
\title{Download the official OSV database archive.}
\usage{
download_osv_database(
  url =
    "https://www.googleapis.com/download/storage/v1/b/osv-vulnerabilities/o/all.zip?alt=media",
  destfile = file.path("osv_db", "all.zip"),
  overwrite = FALSE
)
}
\arguments{
\item{url}{Source URL for the OSV archive.}

\item{destfile}{Destination path for the downloaded archive.}

\item{overwrite}{Whether to overwrite an existing file.}
}
\value{
Absolute path to the downloaded archive.
}
\description{
Download the official OSV database archive.
}
