% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSV-database.R
\name{extract_osv_database}
\alias{extract_osv_database}
\title{Extract OSV archive to a local directory.}
\usage{
extract_osv_database(
  zip_path,
  output_dir = file.path("osv_db", "raw"),
  overwrite = FALSE
)
}
\arguments{
\item{zip_path}{Path to OSV `all.zip`.}

\item{output_dir}{Directory for extracted JSON files.}

\item{overwrite}{Whether to clear output dir before extraction.}
}
\value{
Absolute path to extraction directory.
}
\description{
Extract OSV archive to a local directory.
}
