% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syft-sbom.R
\name{generate_syft_cyclonedx_sbom}
\alias{generate_syft_cyclonedx_sbom}
\title{Generate CycloneDX JSON SBOM using Syft.}
\usage{
generate_syft_cyclonedx_sbom(
  target_path,
  syft_path = "syft.exe",
  output_file = tempfile(pattern = "sbom-", fileext = ".cyclonedx.json"),
  syft_timeout_sec = 0L,
  syft_excludes = NULL
)
}
\arguments{
\item{target_path}{Path to directory/file/image for scanning.}

\item{syft_path}{Path to `syft` executable.}

\item{output_file}{Output path for CycloneDX JSON SBOM.}

\item{syft_timeout_sec}{Timeout for syft process in seconds (`0` means no timeout).}

\item{syft_excludes}{Character vector of syft `--exclude` patterns.}
}
\value{
Absolute path to generated SBOM JSON.
}
\description{
Generate CycloneDX JSON SBOM using Syft.
}
