% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP-OpenAI.R
\name{mcp_tools_to_openai}
\alias{mcp_tools_to_openai}
\title{Convert several MCP-like tools to OpenAI-compatible schemas.}
\usage{
mcp_tools_to_openai(tools)
}
\arguments{
\item{tools}{List of tools created by `mcp_tool()`.}
}
\value{
List of OpenAI-compatible function tool schemas.
}
\description{
Convert several MCP-like tools to OpenAI-compatible schemas.
}
