% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syft-sbom.R
\name{parse_cyclonedx_sbom}
\alias{parse_cyclonedx_sbom}
\title{Parse CycloneDX JSON SBOM and return a flat component table.}
\usage{
parse_cyclonedx_sbom(sbom_file)
}
\arguments{
\item{sbom_file}{Path to CycloneDX JSON.}
}
\value{
Data frame with SBOM components.
}
\description{
Parse CycloneDX JSON SBOM and return a flat component table.
}
