% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Syft-sbom.R
\name{scan_sbom_components_with_osv}
\alias{scan_sbom_components_with_osv}
\title{Scan CycloneDX components against OSV database.}
\usage{
scan_sbom_components_with_osv(components, osv_db, include_uncertain = TRUE)
}
\arguments{
\item{components}{Data frame from `parse_cyclonedx_sbom()`.}

\item{osv_db}{Prepared OSV DB object or path.}

\item{include_uncertain}{Include uncertain version matches.}
}
\value{
Data frame with vulnerabilities for SBOM components.
}
\description{
Scan CycloneDX components against OSV database.
}
