% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSV-search.R
\name{search_osv_vulnerabilities}
\alias{search_osv_vulnerabilities}
\title{Search vulnerabilities in OSV database by purl and version.}
\usage{
search_osv_vulnerabilities(
  osv_db,
  purl,
  version = NULL,
  include_uncertain = TRUE,
  max_results = Inf
)
}
\arguments{
\item{osv_db}{Mongo OSV DB object from `load_osv_mongo_database()`.}

\item{purl}{Package URL (`pkg:` format).}

\item{version}{Package version (optional). If missing, returns potential matches.}

\item{include_uncertain}{Include records where version comparison is inconclusive.}

\item{max_results}{Max number of rows in result.}
}
\value{
Data frame with matched vulnerabilities.
}
\description{
MongoDB backend only.
}
