% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSV-mongo.R
\name{search_osv_vulnerabilities_mongo}
\alias{search_osv_vulnerabilities_mongo}
\title{Search vulnerabilities in MongoDB-backed OSV database by purl and version.}
\usage{
search_osv_vulnerabilities_mongo(
  osv_db,
  purl,
  version = NULL,
  include_uncertain = TRUE,
  max_results = Inf
)
}
\arguments{
\item{osv_db}{Mongo OSV object returned by `load_osv_mongo_database()`.}

\item{purl}{Package URL (`pkg:` format).}

\item{version}{Package version (optional).}

\item{include_uncertain}{Include records where version comparison is inconclusive.}

\item{max_results}{Max number of rows in result.}
}
\value{
Data frame with matched vulnerabilities.
}
\description{
Search vulnerabilities in MongoDB-backed OSV database by purl and version.
}
